api/pyproject.toml
[project]
name = "insani-api"
version = "0.0.4"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.5",
  "psycopg[binary]==3.2.1",
  "opensearch-py==2.6.0",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "numpy>=1.23.5,<2.0.0",  # Compatible with unstructured and opencv
  "pymupdf==1.24.10",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "pillow==10.4.0",
  "pytesseract==0.3.13",
  "opencv-python-headless==4.10.0.84",
  "tiktoken==0.7.0",
  "unstructured[pdf]==0.15.13",
  "beautifulsoup4==4.12.3",
  "lxml==5.3.0",
]

[tool.setuptools]
packages = ["db", "search", "llm", "ingest", "core", "qa"]

